Backward compatibility (#3)

+++ app/src/json/json_parser.c
@@ -214,7 +214,7 @@ parser_error_t json_parse(parsed_json_t *parsed_json, const char *buffer, uint16
     parsed_json->buffer = buffer;
     parsed_json->bufferLen = bufferLen;
 
-    int32_t num_tokens =
+    int16_t num_tokens =
         jsmn_parse(&parser, parsed_json->buffer, parsed_json->bufferLen, parsed_json->tokens, MAX_NUMBER_OF_TOKENS);
 
     ZEMU_LOGF(35, "num_tokens: %d\n", num_tokens);
